Split ScatterProfile module into profile package
