+++ silx/gui/plot/tools/profile/_BaseProfileToolBar.py
@@ -22,33 +22,27 @@
 # THE SOFTWARE.
 #
 # ###########################################################################*/
-"""This module profile tools for scatter plots.
-"""
+"""This module provides the base class for profile toolbars."""
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "30/05/2018"
+__date__ = "07/06/2018"
 
 
 import logging
-import threading
-import time
 import weakref
 
 import numpy
-from scipy.interpolate import LinearNDInterpolator
 
-from ....utils.weakref import WeakMethodProxy
-from ... import qt, icons, colors
-from .. import PlotWidget, items
-from ..ProfileMainWindow import ProfileMainWindow
-from .roi import RegionOfInterestManager
+from silx.utils.weakref import WeakMethodProxy
+from silx.gui import qt, icons, colors
+from silx.gui.plot import PlotWidget, items
+from silx.gui.plot.ProfileMainWindow import ProfileMainWindow
+from silx.gui.plot.tools.roi import RegionOfInterestManager
 
 
 _logger = logging.getLogger(__name__)
 
-
-# TODO support log scale
 
 class _BaseProfileToolBar(qt.QToolBar):
     """Base class for QToolBar plot profiling tools
@@ -413,295 +407,3 @@ class _BaseProfileToolBar(qt.QToolBar):
             roiManager.clearRegionOfInterests()
 
         self._setProfile(profile=None, title='')
-
-
-class _InterpolatorInitThread(qt.QThread):
-    """Thread building a scatter interpolator
-
-    This works in greedy mode in that the signal is only emitted
-    when no other request is pending
-
-    :param QObject parent: See QObject
-    """
-
-    sigInterpolatorReady = qt.Signal(object)
-    """Signal emitted whenever an interpolator is ready
-
-    It provides a 3-tuple (points, values, interpolator)
-    """
-
-    def __init__(self, parent=None):
-        super(_InterpolatorInitThread, self).__init__(parent)
-        self._lock = threading.RLock()
-        self._pendingData = None
-
-    def request(self, points, values):
-        """Request new initialisation of interpolator
-
-        :param numpy.ndarray points: Point coordinates (N, D)
-        :param numpy.ndarray values: Values the N points (1D array)
-        """
-        with self._lock:
-            # Possibly replace already pending data
-            self._pendingData = points, values
-
-        if not self.isRunning():
-            self.start()
-
-    def cancel(self):
-        """Cancel any running/pending requests"""
-        with self._lock:
-            self._pendingData = 'cancelled'
-
-    def run(self):
-        """Run the init of the scatter interpolator"""
-        while True:
-            with self._lock:
-                data = self._pendingData
-                self._pendingData = None
-
-            if data in (None, 'cancelled'):
-                return
-
-            points, values = data
-
-            startTime = time.time()
-            try:
-                interpolator = LinearNDInterpolator(points, values)
-            except:
-                _logger.warning(
-                    "Cannot initialise scatter profile interpolator")
-            else:
-                with self._lock:
-                    data = self._pendingData
-
-                if data is not None:  # Break point
-                    _logger.info('Interpolator discarded after %f s',
-                                 time.time() - startTime)
-                else:
-                    # First call takes a while, do it here
-                    interpolator([(0., 0.)])
-
-                    with self._lock:
-                        data = self._pendingData
-
-                    if data is not None:
-                        _logger.info('Interpolator discarded after %f s',
-                                     time.time() - startTime)
-                    else:
-                        # No other processing requested: emit the signal
-                        _logger.info("Interpolator initialised in %f s",
-                                     time.time() - startTime)
-                        self.sigInterpolatorReady.emit(
-                            (points, values, interpolator))
-
-
-class ScatterProfileToolBar(_BaseProfileToolBar):
-    """QToolBar providing scatter plot profiling tools
-
-    :param parent: See :class:`QToolBar`.
-    :param plot: :class:`PlotWindow` instance on which to operate.
-    :param str title: See :class:`QToolBar`.
-    """
-
-    def __init__(self, parent=None, plot=None, title='Scatter Profile'):
-        super(ScatterProfileToolBar, self).__init__(parent, plot, title)
-
-        self.__nPoints = 1024
-        self.__interpolator = None
-        self.__interpolatorCache = None  # points, values, interpolator
-
-        self.__initThread = _InterpolatorInitThread(self)
-        self.__initThread.sigInterpolatorReady.connect(
-            self.__interpolatorReady)
-
-        roiManager = self._getRoiManager()
-        if roiManager is None:
-            _logger.error(
-                "Error during scatter profile toolbar initialisation")
-        else:
-            roiManager.sigInteractiveModeStarted.connect(
-                self.__interactionStarted)
-            roiManager.sigInteractiveModeFinished.connect(
-                self.__interactionFinished)
-            if roiManager.isStarted():
-                self.__interactionStarted(roiManager.getRegionOfInterestKind())
-
-    def __interactionStarted(self, kind):
-        """Handle start of ROI interaction"""
-        plot = self.getPlotWidget()
-        if plot is None:
-            return
-
-        plot.sigActiveScatterChanged.connect(self.__activeScatterChanged)
-
-        scatter = plot._getActiveItem(kind='scatter')
-        legend = None if scatter is None else scatter.getLegend()
-        self.__activeScatterChanged(None, legend)
-
-    def __interactionFinished(self, rois):
-        """Handle end of ROI interaction"""
-        plot = self.getPlotWidget()
-        if plot is None:
-            return
-
-        plot.sigActiveScatterChanged.disconnect(self.__activeScatterChanged)
-
-        scatter = plot._getActiveItem(kind='scatter')
-        legend = None if scatter is None else scatter.getLegend()
-        self.__activeScatterChanged(legend, None)
-
-    def __activeScatterChanged(self, previous, legend):
-        """Handle change of active scatter
-
-        :param Union[str,None] previous:
-        :param Union[str,None] legend:
-        """
-        self.__initThread.cancel()
-
-        # Reset interpolator
-        self.__interpolator = None
-
-        plot = self.getPlotWidget()
-        if plot is None:
-            _logger.error("Associated PlotWidget no longer exists")
-
-        else:
-            if previous is not None:  # Disconnect signal
-                scatter = plot.getScatter(previous)
-                if scatter is not None:
-                    scatter.sigItemChanged.disconnect(
-                        self.__scatterItemChanged)
-
-            if legend is not None:
-                scatter = plot.getScatter(legend)
-                if scatter is None:
-                    _logger.error("Cannot retrieve active scatter")
-
-                else:
-                    scatter.sigItemChanged.connect(self.__scatterItemChanged)
-                    points = numpy.transpose(numpy.array((
-                        scatter.getXData(copy=False),
-                        scatter.getYData(copy=False))))
-                    values = scatter.getValueData(copy=False)
-
-                    self.__updateInterpolator(points, values)
-
-        # Refresh profile
-        self.updateProfile()
-
-    def __scatterItemChanged(self, event):
-        """Handle update of active scatter plot item
-
-        :param ItemChangedType event:
-        """
-        if event == items.ItemChangedType.DATA:
-            self.__interpolator = None
-            scatter = self.sender()
-            if scatter is None:
-                _logger.error("Cannot retrieve updated scatter item")
-
-            else:
-                points = numpy.transpose(numpy.array((
-                    scatter.getXData(copy=False),
-                    scatter.getYData(copy=False))))
-                values = scatter.getValueData(copy=False)
-
-                self.__updateInterpolator(points, values)
-
-    # Handle interpolator init thread
-
-    def __updateInterpolator(self, points, values):
-        """Update used interpolator with new data"""
-        if (self.__interpolatorCache is not None and
-                len(points) == len(self.__interpolatorCache[0]) and
-                numpy.all(numpy.equal(self.__interpolatorCache[0], points)) and
-                numpy.all(numpy.equal(self.__interpolatorCache[1], values))):
-            # Reuse previous interpolator
-            _logger.info(
-                'Scatter changed: Reuse previous interpolator')
-            self.__interpolator = self.__interpolatorCache[2]
-
-        else:
-            # Interpolator needs update: Start background processing
-            _logger.info(
-                'Scatter changed: Rebuild interpolator')
-            self.__interpolator = None
-            self.__interpolatorCache = None
-            self.__initThread.request(points, values)
-
-    def __interpolatorReady(self, data):
-        """Handle end of init interpolator thread"""
-        points, values, interpolator = data
-        self.__interpolator = interpolator
-        self.__interpolatorCache = None if interpolator is None else data
-        self.updateProfile()
-
-    # Number of points
-
-    def getNPoints(self):
-        """Returns the number of points of the profiles
-
-        :rtype: int
-        """
-        return self.__nPoints
-
-    def setNPoints(self, npoints):
-        """Set the number of points of the profiles
-
-        :param int npoints:
-        """
-        npoints = int(npoints)
-        if npoints < 1:
-            raise ValueError("Unsupported number of points: %d" % npoints)
-        else:
-            self.__nPoints = npoints
-
-    # Overridden methods
-
-    def computeProfileTitle(self, x0, y0, x1, y1):
-        """Compute corresponding plot title
-
-        :param float x0: Profile start point X coord
-        :param float y0: Profile start point Y coord
-        :param float x1: Profile end point X coord
-        :param float y1: Profile end point Y coord
-        :return: Title to use
-        :rtype: str
-        """
-        if self.__initThread.isRunning():
-            return 'Pre-processing data...'
-
-        else:
-            return super(ScatterProfileToolBar, self).computeProfileTitle(
-                x0, y0, x1, y1)
-
-    def computeProfile(self, x0, y0, x1, y1):
-        """Compute corresponding profile
-
-        :param float x0: Profile start point X coord
-        :param float y0: Profile start point Y coord
-        :param float x1: Profile end point X coord
-        :param float y1: Profile end point Y coord
-        :return: (x, y) profile data or None
-        """
-        if self.__interpolator is None:
-            return None
-
-        nPoints = self.getNPoints()
-
-        points = numpy.transpose((
-            numpy.linspace(x0, x1, nPoints, endpoint=True),
-            numpy.linspace(y0, y1, nPoints, endpoint=True)))
-
-        if numpy.abs(x1 - x0) > numpy.abs(y1 - y0):
-            xProfile = points[:, 0]
-        else:
-            xProfile = points[:, 1]
-
-        yProfile = self.__interpolator(points)
-
-        if not numpy.any(numpy.isfinite(yProfile)):
-            return None  # Profile outside convex hull
-
-        return xProfile, yProfile
