+++ silx/gui/plot/tools/profile/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # /*##########################################################################
 #
-# Copyright (c) 2016-2018 European Synchrotron Radiation Facility
+# Copyright (c) 2018 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,33 +22,17 @@
 # THE SOFTWARE.
 #
 # ###########################################################################*/
+"""This module provides tools to get profiles on plot data.
+
+It provides:
+
+- :class:`ScatterProfileToolBar`, a QToolBar to handle profile on scatter data
+
+"""
+
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "29/06/2017"
+__date__ = "07/06/2018"
 
 
-from numpy.distutils.misc_util import Configuration
-
-
-def configuration(parent_package='', top_path=None):
-    config = Configuration('plot', parent_package, top_path)
-    config.add_subpackage('_utils')
-    config.add_subpackage('utils')
-    config.add_subpackage('matplotlib')
-    config.add_subpackage('stats')
-    config.add_subpackage('backends')
-    config.add_subpackage('backends.glutils')
-    config.add_subpackage('items')
-    config.add_subpackage('test')
-    config.add_subpackage('tools')
-    config.add_subpackage('tools.profile')
-    config.add_subpackage('tools.test')
-    config.add_subpackage('actions')
-
-    return config
-
-
-if __name__ == "__main__":
-    from numpy.distutils.core import setup
-
-    setup(configuration=configuration)
+from .ScatterProfileToolBar import ScatterProfileToolBar  # noqa
